more code organisation for map creation
