+++ lib/map_creator.rb
@@ -1,0 +1,6 @@
+module MapCreator
+  require_relative 'map'
+  require_relative 'map_creator/create_tree'
+  require_relative 'map_creator/create_nodes'
+  require_relative 'map_creator/create_rooms'
+end
