+++ lib/map_creator/create_tree.rb
@@ -1,0 +1,15 @@
+module MapCreator
+  class CreateTree
+    def initialize(width:, height:)
+      @width, @height = width, height
+    end
+
+    def call
+      tree = Tree.new
+      tree.root = Node.new(x: 0, y: 0, width: @width, height: @height)
+      tree.nodes.push(tree.root)
+
+      tree
+    end
+  end
+end
